pyproject.toml
[project]
name = "project-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.5.2",
    "great-tables>=0.20.0",
    "keyring>=25.7.0",
    "pandas>=3.0.2",
    "plotnine>=0.15.3",
    "polars>=1.40.0",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "pyfixest>=0.50.1",
    "python-dotenv>=1.2.2",
    "seaborn>=0.13.2",
]
